imp: reorganize client_state.rs + clean-ups
