+++ .changelog/unreleased/improvements/1097-detach-client-state-verifier.md
@@ -1,1 +1,3 @@
-- [ibc-clients/ics07-tendermint] Detach client state verifier such that users have a way to utilize custom verifiers ([\#1097](https://github.com/cosmos/ibc-rs/pull/1097))
+- [ibc-client-tendermint] Detach client state verifier such that users have a
+  way to utilize custom verifiers
+  ([\#1097](https://github.com/cosmos/ibc-rs/pull/1097))
