+++ ibc-clients/ics07-tendermint/src/client_state.rs
@@ -10,61 +10,36 @@
 
 use ibc_client_tendermint_types::error::Error;
 use ibc_client_tendermint_types::proto::v1::ClientState as RawTmClientState;
-use ibc_client_tendermint_types::{
-    client_type as tm_client_type, ClientState as ClientStateType,
-    ConsensusState as ConsensusStateType, Header as TmHeader, Misbehaviour as TmMisbehaviour,
-    TENDERMINT_HEADER_TYPE_URL, TENDERMINT_MISBEHAVIOUR_TYPE_URL,
-};
-use ibc_core_client::context::client_state::{
-    ClientStateCommon, ClientStateExecution, ClientStateValidation,
-};
-use ibc_core_client::context::consensus_state::ConsensusState;
-use ibc_core_client::context::{ClientExecutionContext, ClientValidationContext};
-use ibc_core_client::types::error::{ClientError, UpgradeClientError};
-use ibc_core_client::types::{Height, Status};
-use ibc_core_commitment_types::commitment::{
-    CommitmentPrefix, CommitmentProofBytes, CommitmentRoot,
-};
-use ibc_core_commitment_types::merkle::{apply_prefix, MerkleProof};
-use ibc_core_host::types::identifiers::{ClientId, ClientType};
-use ibc_core_host::types::path::{
-    ClientConsensusStatePath, ClientStatePath, Path, UpgradeClientPath,
-};
-use ibc_core_host::ExecutionContext;
+use ibc_client_tendermint_types::ClientState as ClientStateType;
+use ibc_core_client::types::error::ClientError;
 use ibc_primitives::prelude::*;
 use ibc_primitives::proto::{Any, Protobuf};
-use ibc_primitives::ToVec;
 
-use self::misbehaviour::{check_for_misbehaviour_misbehavior, verify_misbehaviour};
-use self::update_client::{
-    check_for_misbehaviour_update_client, prune_oldest_consensus_state, verify_header,
-};
-use super::consensus_state::ConsensusState as TmConsensusState;
-use crate::context::{
-    CommonContext, DefaultVerifier, ExecutionContext as TmExecutionContext, TmVerifier,
-    ValidationContext as TmValidationContext,
-};
+use crate::context::ValidationContext as TmValidationContext;
 
+mod common;
+mod execution;
 mod misbehaviour;
 mod update_client;
+mod validation;
+
+pub use common::*;
+pub use execution::*;
+pub use misbehaviour::*;
+pub use update_client::*;
+pub use validation::*;
 
 /// Newtype wrapper around the `ClientState` type imported from the
 /// `ibc-client-tendermint-types` crate. This wrapper exists so that we can
 /// bypass Rust's orphan rules and implement traits from
 /// `ibc::core::client::context` on the `ClientState` type.
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq, derive_more::From)]
 pub struct ClientState(ClientStateType);
 
 impl ClientState {
     pub fn inner(&self) -> &ClientStateType {
         &self.0
-    }
-}
-
-impl From<ClientStateType> for ClientState {
-    fn from(client_state: ClientStateType) -> Self {
-        Self(client_state)
     }
 }
 
@@ -100,681 +75,6 @@ impl From<ClientState> for Any {
     }
 }
 
-impl ClientStateCommon for ClientState {
-    fn verify_consensus_state(&self, consensus_state: Any) -> Result<(), ClientError> {
-        verify_consensus_state(consensus_state)
-    }
-
-    fn client_type(&self) -> ClientType {
-        tm_client_type()
-    }
-
-    fn latest_height(&self) -> Height {
-        self.0.latest_height
-    }
-
-    fn validate_proof_height(&self, proof_height: Height) -> Result<(), ClientError> {
-        validate_proof_height(self.inner(), proof_height)
-    }
-
-    fn verify_upgrade_client(
-        &self,
-        upgraded_client_state: Any,
-        upgraded_consensus_state: Any,
-        proof_upgrade_client: CommitmentProofBytes,
-        proof_upgrade_consensus_state: CommitmentProofBytes,
-        root: &CommitmentRoot,
-    ) -> Result<(), ClientError> {
-        verify_upgrade_client(
-            self.inner(),
-            upgraded_client_state,
-            upgraded_consensus_state,
-            proof_upgrade_client,
-            proof_upgrade_consensus_state,
-            root,
-        )
-    }
-
-    fn verify_membership(
-        &self,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        path: Path,
-        value: Vec<u8>,
-    ) -> Result<(), ClientError> {
-        verify_membership(self.inner(), prefix, proof, root, path, value)
-    }
-
-    fn verify_non_membership(
-        &self,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        path: Path,
-    ) -> Result<(), ClientError> {
-        verify_non_membership(self.inner(), prefix, proof, root, path)
-    }
-}
-
-/// Verify an `Any` consensus state by attempting to convert it to a `TmConsensusState`.
-/// Also checks whether the converted consensus state's root is present.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateCommon`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn verify_consensus_state(consensus_state: Any) -> Result<(), ClientError> {
-    let tm_consensus_state = TmConsensusState::try_from(consensus_state)?;
-
-    if tm_consensus_state.root().is_empty() {
-        return Err(ClientError::Other {
-            description: "empty commitment root".into(),
-        });
-    };
-
-    Ok(())
-}
-
-/// Validate the given proof height against the client state's latest height, returning
-/// an error if the proof height is greater than the latest height of the client state.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateCommon`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn validate_proof_height(
-    client_state: &ClientStateType,
-    proof_height: Height,
-) -> Result<(), ClientError> {
-    let latest_height = client_state.latest_height;
-
-    if latest_height < proof_height {
-        return Err(ClientError::InvalidProofHeight {
-            latest_height,
-            proof_height,
-        });
-    }
-
-    Ok(())
-}
-
-/// Perform client-specific verifications and check all data in the new
-/// client state to be the same across all valid Tendermint clients for the
-/// new chain.
-///
-/// You can learn more about how to upgrade IBC-connected SDK chains in
-/// [this](https://ibc.cosmos.network/main/ibc/upgrades/quick-guide.html)
-/// guide.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateCommon`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn verify_upgrade_client(
-    client_state: &ClientStateType,
-    upgraded_client_state: Any,
-    upgraded_consensus_state: Any,
-    proof_upgrade_client: CommitmentProofBytes,
-    proof_upgrade_consensus_state: CommitmentProofBytes,
-    root: &CommitmentRoot,
-) -> Result<(), ClientError> {
-    // Make sure that the client type is of Tendermint type `ClientState`
-    let upgraded_tm_client_state = ClientState::try_from(upgraded_client_state.clone())?;
-
-    // Make sure that the consensus type is of Tendermint type `ConsensusState`
-    TmConsensusState::try_from(upgraded_consensus_state.clone())?;
-
-    let latest_height = client_state.latest_height;
-    let upgraded_tm_client_state_height = upgraded_tm_client_state.latest_height();
-
-    // Make sure the latest height of the current client is not greater then
-    // the upgrade height This condition checks both the revision number and
-    // the height
-    if latest_height >= upgraded_tm_client_state_height {
-        Err(UpgradeClientError::LowUpgradeHeight {
-            upgraded_height: latest_height,
-            client_height: upgraded_tm_client_state_height,
-        })?
-    }
-
-    // Check to see if the upgrade path is set
-    let mut upgrade_path = client_state.upgrade_path.clone();
-
-    if upgrade_path.pop().is_none() {
-        return Err(ClientError::ClientSpecific {
-            description: "cannot upgrade client as no upgrade path has been set".to_string(),
-        });
-    };
-
-    let upgrade_path_prefix = CommitmentPrefix::try_from(upgrade_path[0].clone().into_bytes())
-        .map_err(ClientError::InvalidCommitmentProof)?;
-
-    let last_height = latest_height.revision_height();
-
-    // Verify the proof of the upgraded client state
-    verify_membership(
-        client_state,
-        &upgrade_path_prefix,
-        &proof_upgrade_client,
-        root,
-        Path::UpgradeClient(UpgradeClientPath::UpgradedClientState(last_height)),
-        upgraded_client_state.to_vec(),
-    )?;
-
-    // Verify the proof of the upgraded consensus state
-    verify_membership(
-        client_state,
-        &upgrade_path_prefix,
-        &proof_upgrade_consensus_state,
-        root,
-        Path::UpgradeClient(UpgradeClientPath::UpgradedClientConsensusState(last_height)),
-        upgraded_consensus_state.to_vec(),
-    )?;
-
-    Ok(())
-}
-
-/// Verify membership of the given value against the client's merkle proof.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateCommon`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn verify_membership(
-    client_state: &ClientStateType,
-    prefix: &CommitmentPrefix,
-    proof: &CommitmentProofBytes,
-    root: &CommitmentRoot,
-    path: Path,
-    value: Vec<u8>,
-) -> Result<(), ClientError> {
-    let merkle_path = apply_prefix(prefix, vec![path.to_string()]);
-    let merkle_proof = MerkleProof::try_from(proof).map_err(ClientError::InvalidCommitmentProof)?;
-
-    merkle_proof
-        .verify_membership(
-            &client_state.proof_specs,
-            root.clone().into(),
-            merkle_path,
-            value,
-            0,
-        )
-        .map_err(ClientError::Ics23Verification)
-}
-
-/// Verify that the given value does not belong in the client's merkle proof.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateCommon`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn verify_non_membership(
-    client_state: &ClientStateType,
-    prefix: &CommitmentPrefix,
-    proof: &CommitmentProofBytes,
-    root: &CommitmentRoot,
-    path: Path,
-) -> Result<(), ClientError> {
-    let merkle_path = apply_prefix(prefix, vec![path.to_string()]);
-    let merkle_proof = MerkleProof::try_from(proof).map_err(ClientError::InvalidCommitmentProof)?;
-
-    merkle_proof
-        .verify_non_membership(&client_state.proof_specs, root.clone().into(), merkle_path)
-        .map_err(ClientError::Ics23Verification)
-}
-
-impl<V> ClientStateValidation<V> for ClientState
-where
-    V: ClientValidationContext + TmValidationContext,
-    V::AnyConsensusState: TryInto<TmConsensusState>,
-    ClientError: From<<V::AnyConsensusState as TryInto<TmConsensusState>>::Error>,
-{
-    /// The default verification logic exposed by ibc-rs simply delegates to a
-    /// standalone `verify_client_message` function. This is to make it as simple
-    /// as possible for those who merely need the `DefaultVerifier` behaviour, as
-    /// well as those who require custom verification logic.
-    fn verify_client_message(
-        &self,
-        ctx: &V,
-        client_id: &ClientId,
-        client_message: Any,
-    ) -> Result<(), ClientError> {
-        verify_client_message(
-            self.inner(),
-            ctx,
-            client_id,
-            client_message,
-            &DefaultVerifier,
-        )
-    }
-
-    fn check_for_misbehaviour(
-        &self,
-        ctx: &V,
-        client_id: &ClientId,
-        client_message: Any,
-    ) -> Result<bool, ClientError> {
-        check_for_misbehaviour(self.inner(), ctx, client_id, client_message)
-    }
-
-    fn status(&self, ctx: &V, client_id: &ClientId) -> Result<Status, ClientError> {
-        client_state_status(self.inner(), ctx, client_id)
-    }
-}
-
-/// Verify the client message as part of the client state validation process.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateValidation`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible. It mostly adheres to
-/// the same signature as the `ClientStateValidation::verify_client_message`
-/// function, except for an additional `verifier` parameter that allows users
-/// who require custom verification logic to easily pass in their own verifier
-/// implementation.
-pub fn verify_client_message<V>(
-    client_state: &ClientStateType,
-    ctx: &V,
-    client_id: &ClientId,
-    client_message: Any,
-    verifier: &impl TmVerifier,
-) -> Result<(), ClientError>
-where
-    V: ClientValidationContext + TmValidationContext,
-    V::AnyConsensusState: TryInto<TmConsensusState>,
-    ClientError: From<<V::AnyConsensusState as TryInto<TmConsensusState>>::Error>,
-{
-    match client_message.type_url.as_str() {
-        TENDERMINT_HEADER_TYPE_URL => {
-            let header = TmHeader::try_from(client_message)?;
-            verify_header(client_state, ctx, client_id, &header, verifier)
-        }
-        TENDERMINT_MISBEHAVIOUR_TYPE_URL => {
-            let misbehaviour = TmMisbehaviour::try_from(client_message)?;
-            verify_misbehaviour(client_state, ctx, client_id, &misbehaviour, verifier)
-        }
-        _ => Err(ClientError::InvalidUpdateClientMessage),
-    }
-}
-
-/// Check for misbehaviour on the client state as part of the client state
-/// validation process.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateValidation`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-///
-/// This method covers the following cases:
-///
-/// 1 - fork:
-/// Assumes at least one consensus state before the fork point exists. Let
-/// existing consensus states on chain B be: [Sn,.., Sf, Sf-1, S0] with
-/// `Sf-1` being the most recent state before fork. Chain A is queried for a
-/// header `Hf'` at `Sf.height` and if it is different than the `Hf` in the
-/// event for the client update (the one that has generated `Sf` on chain),
-/// then the two headers are included in the evidence and submitted. Note
-/// that in this case the headers are different but have the same height.
-///
-/// 2 - BFT time violation for unavailable header (a.k.a. Future Lunatic
-/// Attack or FLA):
-/// Some header with a height that is higher than the latest height on A has
-/// been accepted and a consensus state was created on B. Note that this
-/// implies that the timestamp of this header must be within the
-/// `clock_drift` of the client. Assume the client on B has been updated
-/// with `h2`(not present on/ produced by chain A) and it has a timestamp of
-/// `t2` that is at most `clock_drift` in the future. Then the latest header
-/// from A is fetched, let it be `h1`, with a timestamp of `t1`. If `t1 >=
-/// t2` then evidence of misbehavior is submitted to A.
-///
-/// 3 - BFT time violation for existing headers:
-/// Ensure that consensus state times are monotonically increasing with
-/// height.
-pub fn check_for_misbehaviour<V>(
-    client_state: &ClientStateType,
-    ctx: &V,
-    client_id: &ClientId,
-    client_message: Any,
-) -> Result<bool, ClientError>
-where
-    V: ClientValidationContext + TmValidationContext,
-    V::AnyConsensusState: TryInto<TmConsensusState>,
-    ClientError: From<<V::AnyConsensusState as TryInto<TmConsensusState>>::Error>,
-{
-    match client_message.type_url.as_str() {
-        TENDERMINT_HEADER_TYPE_URL => {
-            let header = TmHeader::try_from(client_message)?;
-            check_for_misbehaviour_update_client(client_state, ctx, client_id, header)
-        }
-        TENDERMINT_MISBEHAVIOUR_TYPE_URL => {
-            let misbehaviour = TmMisbehaviour::try_from(client_message)?;
-            check_for_misbehaviour_misbehavior(&misbehaviour)
-        }
-        _ => Err(ClientError::InvalidUpdateClientMessage),
-    }
-}
-
-/// Query the status of the client state.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateValidation`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn client_state_status<V>(
-    client_state: &ClientStateType,
-    ctx: &V,
-    client_id: &ClientId,
-) -> Result<Status, ClientError>
-where
-    V: ClientValidationContext + TmValidationContext,
-    V::AnyConsensusState: TryInto<TmConsensusState>,
-    ClientError: From<<V::AnyConsensusState as TryInto<TmConsensusState>>::Error>,
-{
-    if client_state.is_frozen() {
-        return Ok(Status::Frozen);
-    }
-
-    let latest_consensus_state: TmConsensusState = {
-        let any_latest_consensus_state = match ctx.consensus_state(&ClientConsensusStatePath::new(
-            client_id.clone(),
-            client_state.latest_height.revision_number(),
-            client_state.latest_height.revision_height(),
-        )) {
-            Ok(cs) => cs,
-            // if the client state does not have an associated consensus state for its latest height
-            // then it must be expired
-            Err(_) => return Ok(Status::Expired),
-        };
-
-        any_latest_consensus_state.try_into()?
-    };
-
-    // Note: if the `duration_since()` is `None`, indicating that the latest
-    // consensus state is in the future, then we don't consider the client
-    // to be expired.
-    let now = ctx.host_timestamp()?;
-
-    if let Some(elapsed_since_latest_consensus_state) =
-        now.duration_since(&latest_consensus_state.timestamp().into())
-    {
-        if elapsed_since_latest_consensus_state > client_state.trusting_period {
-            return Ok(Status::Expired);
-        }
-    }
-
-    Ok(Status::Active)
-}
-
-impl<E> ClientStateExecution<E> for ClientState
-where
-    E: TmExecutionContext + ExecutionContext,
-    <E as ClientExecutionContext>::AnyClientState: From<ClientStateType>,
-    <E as ClientExecutionContext>::AnyConsensusState: From<ConsensusStateType>,
-{
-    fn initialise(
-        &self,
-        ctx: &mut E,
-        client_id: &ClientId,
-        consensus_state: Any,
-    ) -> Result<(), ClientError> {
-        initialise_client_state(self.inner(), ctx, client_id, consensus_state)
-    }
-
-    fn update_state(
-        &self,
-        ctx: &mut E,
-        client_id: &ClientId,
-        header: Any,
-    ) -> Result<Vec<Height>, ClientError> {
-        update_client_state(self.inner(), ctx, client_id, header)
-    }
-
-    fn update_state_on_misbehaviour(
-        &self,
-        ctx: &mut E,
-        client_id: &ClientId,
-        client_message: Any,
-    ) -> Result<(), ClientError> {
-        update_client_state_on_misbehaviour(self.inner(), ctx, client_id, client_message)
-    }
-
-    fn update_state_on_upgrade(
-        &self,
-        ctx: &mut E,
-        client_id: &ClientId,
-        upgraded_client_state: Any,
-        upgraded_consensus_state: Any,
-    ) -> Result<Height, ClientError> {
-        update_client_state_on_upgrade(
-            self.inner(),
-            ctx,
-            client_id,
-            upgraded_client_state,
-            upgraded_consensus_state,
-        )
-    }
-}
-
-/// Seed the host store with initial client and consensus states.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateExecution`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn initialise_client_state<E>(
-    client_state: &ClientStateType,
-    ctx: &mut E,
-    client_id: &ClientId,
-    consensus_state: Any,
-) -> Result<(), ClientError>
-where
-    E: TmExecutionContext + ExecutionContext,
-    <E as ClientExecutionContext>::AnyClientState: From<ClientStateType>,
-    <E as ClientExecutionContext>::AnyConsensusState: From<ConsensusStateType>,
-{
-    let host_timestamp = CommonContext::host_timestamp(ctx)?;
-    let host_height = CommonContext::host_height(ctx)?;
-
-    let tm_consensus_state = ConsensusStateType::try_from(consensus_state)?;
-
-    ctx.store_client_state(
-        ClientStatePath::new(client_id.clone()),
-        client_state.clone().into(),
-    )?;
-    ctx.store_consensus_state(
-        ClientConsensusStatePath::new(
-            client_id.clone(),
-            client_state.latest_height.revision_number(),
-            client_state.latest_height.revision_height(),
-        ),
-        tm_consensus_state.into(),
-    )?;
-
-    ctx.store_update_meta(
-        client_id.clone(),
-        client_state.latest_height,
-        host_timestamp,
-        host_height,
-    )?;
-
-    Ok(())
-}
-
-/// Update the host store with a new client state, pruning old states from the
-/// store if need be.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateExecution`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn update_client_state<E>(
-    client_state: &ClientStateType,
-    ctx: &mut E,
-    client_id: &ClientId,
-    header: Any,
-) -> Result<Vec<Height>, ClientError>
-where
-    E: TmExecutionContext + ExecutionContext,
-    <E as ClientExecutionContext>::AnyClientState: From<ClientStateType>,
-    <E as ClientExecutionContext>::AnyConsensusState: From<ConsensusStateType>,
-{
-    let header = TmHeader::try_from(header)?;
-    let header_height = header.height();
-
-    prune_oldest_consensus_state(client_state, ctx, client_id)?;
-
-    let maybe_existing_consensus_state = {
-        let path_at_header_height = ClientConsensusStatePath::new(
-            client_id.clone(),
-            header_height.revision_number(),
-            header_height.revision_height(),
-        );
-
-        CommonContext::consensus_state(ctx, &path_at_header_height).ok()
-    };
-
-    if maybe_existing_consensus_state.is_some() {
-        // if we already had the header installed by a previous relayer
-        // then this is a no-op.
-        //
-        // Do nothing.
-    } else {
-        let host_timestamp = CommonContext::host_timestamp(ctx)?;
-        let host_height = CommonContext::host_height(ctx)?;
-
-        let new_consensus_state = ConsensusStateType::from(header.clone());
-        let new_client_state = client_state.clone().with_header(header)?;
-
-        ctx.store_consensus_state(
-            ClientConsensusStatePath::new(
-                client_id.clone(),
-                header_height.revision_number(),
-                header_height.revision_height(),
-            ),
-            new_consensus_state.into(),
-        )?;
-        ctx.store_client_state(
-            ClientStatePath::new(client_id.clone()),
-            new_client_state.into(),
-        )?;
-        ctx.store_update_meta(
-            client_id.clone(),
-            header_height,
-            host_timestamp,
-            host_height,
-        )?;
-    }
-
-    Ok(vec![header_height])
-}
-
-/// Commit a frozen client state, which was frozen as a result of having exhibited
-/// misbehaviour, to the store.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateExecution`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn update_client_state_on_misbehaviour<E>(
-    client_state: &ClientStateType,
-    ctx: &mut E,
-    client_id: &ClientId,
-    _client_message: Any,
-) -> Result<(), ClientError>
-where
-    E: TmExecutionContext + ExecutionContext,
-    <E as ClientExecutionContext>::AnyClientState: From<ClientStateType>,
-    <E as ClientExecutionContext>::AnyConsensusState: From<ConsensusStateType>,
-{
-    // NOTE: frozen height is  set to `Height {revision_height: 0,
-    // revision_number: 1}` and it is the same for all misbehaviour. This
-    // aligns with the
-    // [`ibc-go`](https://github.com/cosmos/ibc-go/blob/0e3f428e66d6fc0fc6b10d2f3c658aaa5000daf7/modules/light-clients/07-tendermint/misbehaviour.go#L18-L19)
-    // implementation.
-    let frozen_client_state = client_state.clone().with_frozen_height(Height::min(0));
-
-    ctx.store_client_state(
-        ClientStatePath::new(client_id.clone()),
-        frozen_client_state.into(),
-    )?;
-
-    Ok(())
-}
-
-/// Commit the new client state and consensus state to the store.
-///
-/// Note that this function is typically implemented as part of the
-/// [`ClientStateExecution`] trait, but has been made a standalone function
-/// in order to make the ClientState APIs more flexible.
-pub fn update_client_state_on_upgrade<E>(
-    client_state: &ClientStateType,
-    ctx: &mut E,
-    client_id: &ClientId,
-    upgraded_client_state: Any,
-    upgraded_consensus_state: Any,
-) -> Result<Height, ClientError>
-where
-    E: TmExecutionContext + ExecutionContext,
-    <E as ClientExecutionContext>::AnyClientState: From<ClientStateType>,
-    <E as ClientExecutionContext>::AnyConsensusState: From<ConsensusStateType>,
-{
-    let mut upgraded_tm_client_state = ClientState::try_from(upgraded_client_state)?;
-    let upgraded_tm_cons_state = TmConsensusState::try_from(upgraded_consensus_state)?;
-
-    upgraded_tm_client_state.0.zero_custom_fields();
-
-    // Construct new client state and consensus state relayer chosen client
-    // parameters are ignored. All chain-chosen parameters come from
-    // committed client, all client-chosen parameters come from current
-    // client.
-    let new_client_state = ClientStateType::new(
-        upgraded_tm_client_state.0.chain_id,
-        client_state.trust_level,
-        client_state.trusting_period,
-        upgraded_tm_client_state.0.unbonding_period,
-        client_state.max_clock_drift,
-        upgraded_tm_client_state.0.latest_height,
-        upgraded_tm_client_state.0.proof_specs,
-        upgraded_tm_client_state.0.upgrade_path,
-        client_state.allow_update,
-    )?;
-
-    // The new consensus state is merely used as a trusted kernel against
-    // which headers on the new chain can be verified. The root is just a
-    // stand-in sentinel value as it cannot be known in advance, thus no
-    // proof verification will pass. The timestamp and the
-    // NextValidatorsHash of the consensus state is the blocktime and
-    // NextValidatorsHash of the last block committed by the old chain. This
-    // will allow the first block of the new chain to be verified against
-    // the last validators of the old chain so long as it is submitted
-    // within the TrustingPeriod of this client.
-    // NOTE: We do not set processed time for this consensus state since
-    // this consensus state should not be used for packet verification as
-    // the root is empty. The next consensus state submitted using update
-    // will be usable for packet-verification.
-    let sentinel_root = "sentinel_root".as_bytes().to_vec();
-    let new_consensus_state = ConsensusStateType::new(
-        sentinel_root.into(),
-        upgraded_tm_cons_state.timestamp(),
-        upgraded_tm_cons_state.next_validators_hash(),
-    );
-
-    let latest_height = new_client_state.latest_height;
-    let host_timestamp = CommonContext::host_timestamp(ctx)?;
-    let host_height = CommonContext::host_height(ctx)?;
-
-    ctx.store_client_state(
-        ClientStatePath::new(client_id.clone()),
-        new_client_state.into(),
-    )?;
-    ctx.store_consensus_state(
-        ClientConsensusStatePath::new(
-            client_id.clone(),
-            latest_height.revision_number(),
-            latest_height.revision_height(),
-        ),
-        new_consensus_state.into(),
-    )?;
-    ctx.store_update_meta(
-        client_id.clone(),
-        latest_height,
-        host_timestamp,
-        host_height,
-    )?;
-
-    Ok(latest_height)
-}
-
 #[cfg(test)]
 mod tests {
     use core::time::Duration;
@@ -782,8 +82,10 @@ mod tests {
     use ibc_client_tendermint_types::{
         AllowUpdate, ClientState as ClientStateType, TrustThreshold,
     };
+    use ibc_core_client::types::Height;
     use ibc_core_commitment_types::specs::ProofSpecs;
     use ibc_core_host::types::identifiers::ChainId;
+    use tests::common::validate_proof_height;
 
     use super::*;
