+++ ibc-core/ics24-host/cosmos/src/validate_self_client.rs
@@ -1,6 +1,7 @@
 use core::time::Duration;
 
 use ibc_client_tendermint::client_state::ClientState;
+use ibc_core_client_context::client_state::ClientStateCommon;
 use ibc_core_client_types::error::ClientError;
 use ibc_core_client_types::Height;
 use ibc_core_commitment_types::specs::ProofSpecs;
@@ -49,7 +50,7 @@ pub trait ValidateSelfClientContext {
             ));
         }
 
-        let latest_height = tm_client_state.inner().latest_height;
+        let latest_height = tm_client_state.latest_height();
         let self_revision_number = self_chain_id.revision_number();
         if self_revision_number != latest_height.revision_number() {
             return Err(ContextError::ConnectionError(
@@ -68,7 +69,7 @@ pub trait ValidateSelfClientContext {
                 ConnectionError::InvalidClientState {
                     reason: format!(
                         "client has latest height {} greater than or equal to chain height {}",
-                        tm_client_state.inner().latest_height,
+                        latest_height,
                         self.host_current_height()
                     ),
                 },
